timeseries tests, tox fixes
