+++ redisplus/client.py
@@ -1,4 +1,4 @@
-from typing import Dict, Optional
+from typing import Optional
 from redis.client import Redis
