+++ redisplus/ts/client.py
@@ -68,11 +68,6 @@ class Client(CommandMixin, AbstractFeature, object):
         for k in MODULE_CALLBACKS:
             self.client.set_response_callback(k, MODULE_CALLBACKS[k])
 
-    @property
-    def client(self):
-        """Get the client."""
-        return self.CLIENT
-
     @staticmethod
     def appendUncompressed(params, uncompressed):
         """Append UNCOMPRESSED tag to params."""
@@ -164,9 +159,22 @@ class Client(CommandMixin, AbstractFeature, object):
         if min_value is not None and max_value is not None:
             params.extend(["FILTER_BY_VALUE", min_value, max_value])
 
-    def execute_command(self, *args, **kwargs):
-        """Execute redis command."""
-        return self.client.execute_command(*args, **kwargs)
+    def pipeline(self, transaction=True, shard_hint=None):
+        """
+        Return a new pipeline object that can queue multiple commands for later execution.
+
+        ``transaction`` indicates whether all commands should be executed atomically.
+        Apart from making a group of operations atomic, pipelines are useful for reducing
+        the back-and-forth overhead between the client and server.
+        Overridden in order to provide the right client through the pipeline.
+        """
+        p = Pipeline(
+            connection_pool=self.client.connection_pool,
+            response_callbacks=self.client.response_callbacks,
+            transaction=transaction,
+            shard_hint=shard_hint,
+        )
+        return p
 
 
 class Pipeline(Pipeline, Client):
