+++ tox.ini
@@ -14,14 +14,14 @@ ports =
 [flake8]
 max-complexity = 10
 filename = ./redisplus/**
-ignore = W292, E501
+ignore = W292, E501,D104,D100
 exclude = conftest.py
 show-source = true
 
 [testenv:linters]
 skipsdist = true
 commands =
-    flake8 redisplus
+    flake8 redisplus/*
     black --check --diff redisplus
     vulture redisplus --min-confidence 80
     bandit -r redisplus
